fixed bug 2904
git-svn-id: http://svn.osafoundation.org/chandler/trunk@5163 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/repository/tests/TestPerfWithRSS.py
@@ -143,7 +143,11 @@ class TestPerfWithRSS(RepositoryTestCase):
 
     def tearDown(self):
         self.rep.close()
-        self.rep.delete()
+        if os.path.exists(self.rep.dbHome):
+            self.rep.delete()
+        else:
+            self.rep.logger.warn("no repository at %s", self.rep.dbHome)
+        
 
 if __name__ == "__main__":
 #    import hotshot
